Split extension messaging Sentry filter
