+++ utils/sentry-client-filters.ts
@@ -24,8 +24,8 @@ export {
   shouldFilterSentryRouteParameterizationError,
   shouldFilterTwitterConfigReferenceError,
 } from "./sentry-client-filters/errors";
+export { shouldFilterBrowserExtensionMessagingConnectionError } from "./sentry-client-filters/extension-messaging";
 export {
-  shouldFilterBrowserExtensionMessagingConnectionError,
   shouldFilterCoinbaseWalletLinkWebSocket1006,
   shouldFilterDisconnectedWalletProviderRejection,
   shouldFilterInjectedProviderProxyStartsWithError,
