+++ utils/sentry-client-filters/app-frame-utils.ts
@@ -18,11 +18,13 @@ import {
 import type {
   SentryClientEvent,
   SentryEventHint,
+  SentryExceptionValue,
   SentryStackFrame,
 } from "./types";
 import {
   getFramePaths,
   getHintExceptionStack,
+  getSerializedExceptionStack,
   isFirstPartyHost,
 } from "./value-utils";
 
@@ -384,6 +386,19 @@ export function hasAppOwnedSourceStackValue(value: string): boolean {
   }
 
   return getStackFramePathCandidates(value).some(isAppOwnedStackPath);
+}
+
+export function hasAppOwnedSourceEvidence(
+  event: SentryClientEvent,
+  value: SentryExceptionValue | undefined,
+  hint?: SentryEventHint
+): boolean {
+  const frames = value?.stacktrace?.frames;
+  return (
+    hasAppOwnedSourceFrame(frames) ||
+    hasAppOwnedSourceStackValue(getHintExceptionStack(hint)) ||
+    hasAppOwnedSourceStackValue(getSerializedExceptionStack(event))
+  );
 }
 
 function getStackFramePathCandidates(value: string): string[] {
